chatbot/requirements.txt
streamlit>=1.20.0
python-dotenv>=1.0.0
pyyaml>=6.0
requests>=2.28.0
